Added more debugging statements.

+++ NodeVersionControl/Operations/Change.cs
@@ -10,12 +10,18 @@
                 newVersion = "v" + newVersion;
 
             if (oldVersion == newVersion)
-                throw new Exception($"Already using version {newVersion} of NodeJS.");
+            {
+                Console.WriteLine($"Already using version {newVersion} of NodeJS.");
+                return;
+            }
 
             string newVersionPath = Path.Combine(Globals.NODE_VERSIONS_DIRECTORY, newVersion);
 
             if (!Directory.Exists(newVersionPath))
-                throw new Exception($"NodeJS version {newVersion} not found.");
+            {
+                Console.WriteLine($"NodeJS version {newVersion} not installed.");
+                return;
+            }
 
             StoreGlobalNodeModules(oldVersion);
 
@@ -33,18 +39,47 @@
         {
             //Delete current NodeJS version
             if (Directory.Exists(Globals.NODE_DIRECTORY))
+            {
+                if (Globals.DEBUG)
+                {
+                    Console.WriteLine($"DEBUG: Removing files in working NodeJS directory {Globals.NODE_DIRECTORY}");
+                }
+
                 SharedMethods.DeleteDirectory(Globals.NODE_DIRECTORY, keepRootFolder: true);
+            }
             else
+            {
                 Directory.CreateDirectory(Globals.NODE_DIRECTORY);
+            }
 
             //Copy over requested NodeJS version
+            if (Globals.DEBUG)
+            {
+                Console.WriteLine($"DEBUG: Copying new NodeJS version from {versionDirectory} into working NodeJS directory.");
+            }
+
             SharedMethods.CopyDirectoryContents(versionDirectory, Globals.NODE_DIRECTORY);
         }
 
         private static void RestoreGlobalNodeModules(string versionDirectory)
         {
-            //Copy over global NPM packages
-            SharedMethods.CopyDirectoryContents(Path.Combine(versionDirectory, Globals.NPM_GLOBALS_STORAGE_FOLDER_NAME), Globals.NPM_GLOBALS_DIRECTORY);
+            string npmGlobalsStorageFolder = Path.Combine(versionDirectory, Globals.NPM_GLOBALS_STORAGE_FOLDER_NAME);
+
+            if (Globals.DEBUG)
+            {
+                Console.WriteLine($"DEBUG: Restoring NPM Global packages from storage folder {npmGlobalsStorageFolder} to {Globals.NPM_GLOBALS_DIRECTORY}");
+            }
+
+            if (Directory.Exists(npmGlobalsStorageFolder))
+            {
+                //Copy over global NPM packages
+                SharedMethods.CopyDirectoryContents(npmGlobalsStorageFolder, Globals.NPM_GLOBALS_DIRECTORY);
+            }
+            else if (Globals.DEBUG)
+            {
+                Console.WriteLine($"DEBUG: No stored NPM global modules found.");
+            }
+            
         }
 
         private static void StoreGlobalNodeModules(string currentVersion)
@@ -54,17 +89,37 @@
             {
                 string versionedNpmGlobalsDirectory = Path.Combine(Globals.NODE_VERSIONS_DIRECTORY, currentVersion, Globals.NPM_GLOBALS_STORAGE_FOLDER_NAME);
 
+                if (Globals.DEBUG)
+                {
+                    Console.WriteLine($"DEBUG: Saving NPM Globals Directory to storage folder. {versionedNpmGlobalsDirectory}");
+                }
+
                 if (!Directory.Exists(versionedNpmGlobalsDirectory))
                 {
                     Directory.CreateDirectory(versionedNpmGlobalsDirectory);
                 }
                 else
                 {
+                    if (Globals.DEBUG)
+                    {
+                        Console.WriteLine($"DEBUG: Version {currentVersion} contains old NPM Global Packages, purging folder {versionedNpmGlobalsDirectory}");
+                    }
+
                     SharedMethods.DeleteDirectory(versionedNpmGlobalsDirectory, keepRootFolder:true);
                 }
 
                 SharedMethods.CopyDirectoryContents(Globals.NPM_GLOBALS_DIRECTORY, versionedNpmGlobalsDirectory);
+
+                if (Globals.DEBUG)
+                {
+                    Console.WriteLine($"DEBUG: Purging NPM Globals Directory for next version.");
+                }
+
                 SharedMethods.DeleteDirectory(Globals.NPM_GLOBALS_DIRECTORY, keepRootFolder: true);
+            }
+            else if(Globals.DEBUG)
+            {
+                Console.WriteLine($"DEBUG: NPM Globals package folder {Globals.NPM_GLOBALS_DIRECTORY} did not exist.");
             }
         }
     }
